Added all post function

+++ app/models/post.server.models.js
@@ -1,10 +1,10 @@
 const db = require('../../database');
 const crypto = (require('crypto'));
 
-const addNewPost = (post,user_id, done) => {
+const addNewPost = (post, user_id, done) => {
     const sql = 'INSERT INTO posts (text, date_published, author_id) VALUES (?, ?, ?)';
 
-    
+
     let values = [post.text, Date.now(), user_id]
 
     db.run(sql, values, function (err) {
@@ -18,9 +18,9 @@ const getSinglePost = (post_id, done) => {
 
     db.get(sql, [post_id], function (err, post_details) {
         if (err) return done(err)
-        if (!post_details){
+        if (!post_details) {
             console.log("No post details")
-             return done(404);
+            return done(404);
         }
 
         const sql = 'SELECT u.user_id, u.first_name, u.last_name, u.username FROM users u, likes l WHERE l.post_id=? AND l.user_id=u.user_id'
@@ -60,30 +60,70 @@ const updatePost = (post_id, new_text, done) => {
         return done(err);
     })
 }
-const deletePost = (post_id, done) => {
-    const sql = "DELETE FROM posts WHERE post_id=?";
-    db.run(sql, [post_id], (err) => {
-        return done(err)
+const deletePost = (post_id, user_id, done) => {
+    const sql = "SELECT post_id, author_id FROM posts WHERE post_id=?"
+    db.get(sql, [post_id], (err, row) => {
+        if (err) return done(err);
+        if (!row) return done(404);
+        if (row.author_id != user_id) {
+            return done(403);
+        }
+        const sql = "DELETE FROM posts WHERE post_id =?"
+        db.run(sql, post_id, (err) => {
+            if(err) return done(err);
+        })
+        return done(null)
     })
+
+
 }
-const addLike = (post_id, done) => {
-    const sql = 'INSERT INTO likes (post_id, user_id) VALUES (?,?)'
-    const values = [post_id, 6]
-    db.run(sql, values, (err) => {
+const addLike = (post_id, user_id, done) => {
+    const sql = 'SELECT post_id FROM posts WHERE post_id=?';
+    
+    //CHECK POST EXISTS
+    db.get(sql, [post_id], function(err, post_details) {
         if (err) return done(err)
-        db.all("SELECT l.post_id FROM likes l WHERE l.post_id=?", [post_id], (err, rows) => {
-            if (err) return done(err)
-            return done(null, rows.length);
+        if (!post_details)
+            return done(404);
+        //adds like
+        const sql = 'INSERT INTO likes (post_id, user_id) VALUES (?,?)'
+        const values = [post_id, user_id]
+        db.run(sql, values, (err) => {
+            if(err){
+                //error code for unique constraint
+                if(err.errno === 19) return done(403)
+    
+                return done(err)
+            }
+            return done(null)
         })
     })
 }
-const removeLike = (post_id, done) => {
-    const sql = 'DELETE FROM likes WHERE post_id=?';
-    db.run(sql, post_id, (err) => {
-        if (err) return done(err)
-        db.all("SELECT l.post_id FROM likes l WHERE l.post_id=?", [post_id], (err, rows) => {
-            if (err) return done(err)
-            return done(null, rows.length);
+const removeLike = (post_id, user_id, done) => {
+    //get post
+
+    const sql = 'SELECT post_id FROM posts WHERE post_id = ?'
+    db.get(sql, [post_id], function(err, row){
+        if (err) return done(err);
+        if (!row) return done(404);
+
+        console.log("POST EXISTS")
+        console.log(row)
+    
+        const sql = 'SELECT l.post_id, l.user_id FROM likes l WHERE l.post_id=? AND l.user_id=?'
+        db.get(sql, [post_id, user_id], (err, row) => {
+            if(err) return done(err)
+            console.log("ROW PRINT")
+            console.log(row)
+            if(!row) return done(403)
+            console.log("HAS LIKED")
+            const sql = 'DELETE FROM likes WHERE post_id=? AND user_id = ?'
+            db.run(sql, [post_id, user_id], (err) => {
+                console.log("DELETE")
+                console.log(err )
+                if (err) return done(err)
+                return done(null)
+            })
         })
     })
 }
